worker/requirements.txt
sqlalchemy==2.0.36
psycopg[binary]==3.2.2
redis==5.0.8
httpx==0.27.2
openai==1.47.0
qdrant-client==1.11.1
pydantic-settings==2.4.0
pypdf==4.3.1
python-docx==1.1.2
beautifulsoup4==4.12.3
numpy==1.26.4
pytest==8.2.2
boto3==1.34.162
loguru==0.7.2
